Added script to generate socio-demogrpahic layers
Code pulls data relating to car ownership, population density, employment density and IMD, generated appropriate bandings and outputs at most granular geospatial level available.
